Start refactor circuit optimization to MLIR passes
Having each of them in a pass will help to debug and analyze each subroutine more easily.
