+++ mlir/transforms/CMakeLists.txt
@@ -1,7 +1,7 @@
 
 set(LIBRARY_NAME quantum-to-llvm-lowering)
 
-file(GLOB SRC *.cpp lowering/*.cpp)
+file(GLOB SRC *.cpp lowering/*.cpp optimizations/*.cpp)
 
 add_subdirectory(optimizations/utils)
